tune knight pst properly
Score of Koivisto Dev vs Koivisto 0.003 speedups: 1216 - 1086 - 1698 [0.516]
...      Koivisto Dev playing White: 640 - 523 - 837  [0.529] 2000
...      Koivisto Dev playing Black: 576 - 563 - 861  [0.503] 2000
...      White vs Black: 1203 - 1099 - 1698  [0.513] 4000
Elo difference: 11.3 +/- 8.2, LOS: 99.7 %, DrawRatio: 42.4 %
4000 of 4000 games finished.

+++ src_files/eval.h
@@ -44,10 +44,10 @@ class Evaluator{
         int paramCount();
         
         #ifdef TUNE_PST
-        double* getTunablePST_MG();
-        double* getTunablePST_EG();
-        double* getTunablePST_MG_grad();
-        double* getTunablePST_EG_grad();
+        float* getTunablePST_MG();
+        float* getTunablePST_EG();
+        float* getTunablePST_MG_grad();
+        float* getTunablePST_EG_grad();
         #endif
     
 };
